+++ posthog-android/src/test/java/com/posthog/android/PostHogFake.kt
@@ -80,6 +80,10 @@ public class PostHogFake : PostHogInterface {
         return null
     }
 
+    override fun getAllFeatureFlags(): List<FeatureFlagResult>? {
+        return null
+    }
+
     override fun flush() {
         flushes++
     }
@@ -164,6 +168,10 @@ public class PostHogFake : PostHogInterface {
     }
 
     override fun distinctId(): String {
+        return ""
+    }
+
+    override fun getDeviceId(): String {
         return ""
     }